Completely rewrote gfx
The last commit suffered from a design issue: handle.id == 0 is an valid
id. Passing temporary structs to functions causes attributes, which are
not set, to be zero.
As of this commit, this is no longer the case.

Additionally, structures and enumerations were moved atop of functions
and devices now can be distinguished.

As a final note, render targets, including the swap chain will probably
be part of the of the 'fn_buffer_binding' structure.
The structure will probably renamed to 'fn_resource_binding', allowing
buffers, textures (read and/or writable) as well as constant buffers to
be bound.

It might make sense to make them a primitive in order to reduce look up
and mapping times.