Completely moved to CMake again; xcb support somewhat complete.
With this commit this project moves back to CMake from Qbs.

While I dearly love Qbs, it lacks features; for instance it lacks
the ability to specify the 'Platform Toolset' on Windows, making it hard
to create good NuGet packages.
Also its Visual Studio generators are broken, causing the build process to
crash.

Regarding xcb support, the window operations such as create, destroy etc.
are complete; the poll event function however does not handle any event besides
the 'closed' event, without providing the window that was closed.

This is to be changed in upcoming commits.