Fixed minor mistake in CMakeLists, created 'common' event handlers
This commit fixes a minor mistake in the 'lib/c/CMakeLists.txt', which caused
the build to fail on macOS, since the require framework 'AppKit' was not linked.

Additionally 'common' event handlers were introduced into the module
'window_common', providing, as the name implies, event handlers or callbacks
to be used by the target implementations when processing an event.