Introducing keymap_gen; trying to fix #12, #6, and #5
The mapping between fundament key codes and platform native key codes was
previously maintained by hand. This led to a lot of errors and a loss of
overview.

To combat this, the mapping is now maintained in an online spreadsheet
https://docs.google.com/spreadsheets/d/1VPQTYbJqPITDeBr0M5-fUwX6cvYu9hFFXQVaXtD9sN0/edit?usp=sharing
which is then processed by the keymap_gen.py script. The script then generates
the source for a function, that translates the native key code to fundaments
key code.

Note, that the script itself is horribly written and just hacked together.